Relocate CompactionJobStatusTestData to core module
