+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/job/CompactionJobStatusFromJobTestData.java
@@ -16,6 +16,7 @@
 package sleeper.compaction.core.job;
 
 import sleeper.core.record.process.status.ProcessRun;
+import sleeper.core.tracker.compaction.job.CompactionJobStatusTestData;
 import sleeper.core.tracker.compaction.job.query.CompactionJobStatus;
 
 import java.time.Instant;
